pyproject.toml
[project]
name = "deepagents"
version = "0.0.3"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and postgres database tools. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-anthropic>=0.1.23",
    "langchain-openai>=0.1.0",
    "langchain>=0.2.14",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
]


[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deepagents"]
[tool.setuptools.package-dir]
"deepagents" = "src/deepagents"

[tool.setuptools.package-data]
"*" = ["py.typed"]
